chore(calculator): rename Level -> LevelRecords
