+++ app/src/Calculator/constants/levelRecords.ts
