+++ app/src/Calculator/dashboard-contents/index.ts
@@ -1,2 +1,2 @@
 export * from './Blackhole';
-export * from './Level';
+export * from './LevelRecords';
